+++ Runtime/Jobs/ITransformJobUpdatable.cs
@@ -1,3 +1,4 @@
+using System;
 using Gilzoide.UpdateManager.Jobs.Internal;
 
 namespace Gilzoide.UpdateManager.Jobs
@@ -12,8 +13,12 @@ namespace Gilzoide.UpdateManager.Jobs
 
     /// <summary>
     /// Alias for <see cref="ITransformJobUpdatable{}"/>.
-    /// Pass <c>BurstUpdateTransformJob&lt;<typeparamref name="TData"/>&gt;</c> as <typeparamref name="TJob"/> to Burst compile your job.
     /// </summary>
+    /// <remarks>
+    ///   Deprecated: use <see cref="ITransformJobUpdatable{}"/> instead and implement
+    ///   <see cref="IBurstUpdateTransformJob{}"/> in job definition to compile jobs with Burst.
+    /// </remarks>
+    [Obsolete("Use ITransformJobUpdatable<> and implement IBurstUpdateTransformJob<> in job definition instead.")]
     public interface ITransformJobUpdatable<TData, TJob> : ITransformJobUpdatable<TData>
         where TData : struct, IUpdateTransformJob
         where TJob : struct, IInternalUpdateTransformJob<TData>