+++ Runtime/Jobs/Internal/UpdateJobOptions.cs
@@ -79,5 +79,13 @@ namespace Gilzoide.UpdateManager.Jobs.Internal
             }
             return managers;
         }
+
+#if HAVE_BURST
+        public static bool GetIsBurstCompiled<TData>()
+        {
+            return typeof(TData).ImplementsGenericInterface(typeof(IBurstUpdateJob<>))
+                || typeof(TData).ImplementsGenericInterface(typeof(IBurstUpdateTransformJob<>));
+        }
+#endif
     }
 }