+++ Runtime/Jobs/UpdateTransformJobManager.cs
@@ -12,8 +12,8 @@ namespace Gilzoide.UpdateManager.Jobs
     /// <br/>
     /// This class runs jobs in parallel, so don't rely on jobs being executed in any order.
     /// </remarks>
-    /// <seealso cref="UnityEngine.Jobs.IJobParallelForTransformExtensions.Schedule"/>
-    /// <seealso cref="UnityEngine.Jobs.IJobParallelForTransformExtensions.ScheduleReadOnly"/>
+    /// <seealso cref="IJobParallelForTransformExtensions.Schedule"/>
+    /// <seealso cref="IJobParallelForTransformExtensions.ScheduleReadOnly"/>
     public class UpdateTransformJobManager<TData> : AUpdateJobManager<TData, ITransformJobUpdatable<TData>, UpdateTransformJobData<TData, ITransformJobUpdatable<TData>>>
         where TData : struct, IUpdateTransformJob
     {
@@ -26,7 +26,22 @@ namespace Gilzoide.UpdateManager.Jobs
 
         protected override JobHandle ScheduleJob(JobHandle dependsOn)
         {
-            var job = new UpdateTransformJob<TData>
+#if HAVE_BURST
+            if (IsJobBurstCompiled)
+            {
+                return Schedule<BurstUpdateTransformJob<TData>>(dependsOn);
+            }
+            else
+#endif
+            {
+                return Schedule<UpdateTransformJob<TData>>(dependsOn);
+            }
+        }
+
+        protected JobHandle Schedule<TJob>(JobHandle dependsOn)
+            where TJob : struct, IInternalUpdateTransformJob<TData>
+        {
+            var job = new TJob
             {
                 Data = _jobData.Data,
             };